pyproject.toml
[project]
name = "bpoptima"
version = "0.1.0"
description = "Async document processing API"
authors = [
    {name = "Tanay Shah"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "slowapi (>=0.1.9,<0.2.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "celery[redis] (>=5.6.3,<6.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "bcrypt (>=3.2.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-env (>=1.6.0,<2.0.0)"
]

[[tool.poetry.packages]]
include = "app"
